Build the leftmost menu with chapters, summaries, characters menu options
